+++ packages/plugin-spoiler/web/src/index.ts
@@ -1,4 +1,6 @@
 export { createSpoilerPlugin } from './createSpoilerPlugin';
 export { default as SpoilerViewer } from './spoiler-viewer';
 export { SPOILER_TYPE } from './types';
-export { pluginSpoiler } from './editor';
+export { default as TextSpoilerButton } from './toolbar/TextSpoilerButton';
+export { default as BlockSpoilerComponent } from './BlockSpoilerComponent';
+export { pluginSpoiler, SpoilerEditorWrapper } from './editor';